Front End 1.01
A few more minor tweaks:

made background light blue

I added some checks to the edit route, if you don’t give it an id it
will alert you and you’ll get sent home. if you give an invalid id
you’ll be told the id is invalid and get sent home.  It took me awhile
to figure out how to interpret the response I was getting sent from the
API - originally I was trying to check to see if err existed but that
didn’t seem to work so eventually I just checked to make sure the
response wasn’t null to see if the id was valid.

I also realized that the backend logic isn’t successfully validating
the model.  I can send it blank data and it returns me a “200” response
(which means success) and it puts a blank listing into the DB.  I added
some checks to the front end to prevent this, but the backend should
also verify this.

I made it so that after a successful new listing is made or an edit is
made successfully, you are taken home. This behavior can be tweaked but
before it wasn’t obvious if the button was working without watching the
console.

I also removed the “app.js” file within the public directory because it
actually wasn’t being used.  All of the controller data/etc is in
maincontroller.js inside public/controllers.