Front/Back V3
Changes:
Username attached to new listings as seller automatically

“your account” link at top which shows you a link of  your own listings
if you have any

“detailed search” allows you to search by only one field, or multiple.
link is underneath the main search bar, I haven’t decided on how the ui
should be laid out.  Detailed search will also not return listings that
belong to yourself, because I figured that made sense.

You may want to drop your old listings documents if you have a bunch
that have null fields.  I was having problems with the filtering based
on seller name and I eventually got it to work after dropping all the
documents and starting over - I don’t know if it was just coincidence
or if it actually did something, but all of a sudden I got it working.

On the detailed search page, each result has a link created which will
take you to a separate page which shows just that listing’s info. This
page will have to be worked on, but this is where I will put the “buy”
button.

The “your account” page is where I will update to show if one of your
listings has a buy offer

I still haven’t updated the front end to reflect the updated listing
model, I was trying to get it working with less info at first, I can
add the extra stuff in now